Create some shared logic and tests for LLM clients
